Split compat.c into smaller units
When linking libmicrodns into a larger binary, the larger binary
(an enclosing VLC plugin) might also be providing similar replacements
for certain functions, like poll and inet_ntop.

If the optional functions are in separate object files, the
linker may choose not to include them if it already has got a
definition from another object file - but if all of them are
bundled up in the same, one can run into multiple definition errors,
if the linker already has chosen to include the other object file,
and then later tries to include libmicrodns' compat.o.

This fixes linking into the VLC plugin when linked with LLD.